[Case note search] Add CORS support to prototypes app (Recidiviz/recidiviz-data#31809)
## Description of the change

The [Same Origin
Policy](https://en.wikipedia.org/wiki/Same-origin_policy) prevents a
frontend at http://xxx.com from fetching resources from an API at
http://yyy.com. To allow this, you need to use CORS, which allows the
server to indicate to a browser that certain resources can be used. It
does so via access-control headers that inform the browser how the
resource can be used. For GET requests these headers are sent in the
response. For non-GET requests the browser must ask the server for the
access-control headers before sending the actual request, and it does so
via a preflight OPTIONS request. (For some reason even though we are
only using GET, I still had to add support for the preflight OPTIONS
request).

I used the same logic used by Insights
[here](https://github.com/Recidiviz/recidiviz-data/blob/99592833b6bcaa18b9b4e48ee64146eec9aa7cbc/recidiviz/case_triage/outliers/outliers_routes.py#L69).

Tested by hacking a button into the recidiviz-dashboard frontend that
calls this backend, and it worked!

Note that at first I tried to use the quart-cors extension, but this
required a newer version of quart, and when I tried to upgrade us (we're
on 17, and the latest is 19) the Pipfile failed to lock. Just mentioning
this here in case we ever need to upgrade quart in the future and run
into issues ...

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 701581efdbbc34884c201abe99374624ba66b428