handle case where new columns aren't found (Recidiviz/recidiviz-data#31830)
## Description of the change

During yesterday's staging deploy, there were new columns added to the
schema that hadn't made it into an export yet, causing 500 errors in the
backend. This applies a quick-and-dirty fix so that this doesn't happen
in prod next week, by just catching the exception and trying again
without the new columns. This should make it into today's release
canddiate so this prevention can happen.

Tested by loading fixtures to my local DB at an older commit (so the new
columns don't exist), and manually sending requests to a local server to
make sure they don't fail.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 20a74c423818eb1496a6ff70b554bd5419a7bc88