[US_MO]: Ingest StateSentenceGroupLength with parole projected dates (Recidiviz/recidiviz-data#31676)
An ingest view and mapping for `StateSentenceGroupLength` with parole
eligibility and release dates.

#### Assumptions/Facts
See the
[issue](https://app.zenhub.com/workspaces/data-platform-61e76c3acfd89e0014d537de/issues/gh/recidiviz/recidiviz-data/31186)
for queries used in exploration
- We track changes over time using the date created `CG_DCR`
- Eligibility Sequence Number does not always start at 1, but is needed
to keep unique records.
- P & P report date was NOT good for the critical date. We instead use
date updated/created
- The parole dates have varied invalid values. Many look to be invalid
data entries. I null them out.
- Parole dates can be NULL after they have been NOT NULL. An example of
this happening is a revocation and additional charge, where the new
sentence may not be parole eligible. This then nulls out parole dates
for the group.

GitOrigin-RevId: d3c51a82b0b3d7cf56c4c06536883515a472130b