[Insights] use server_default instead of default for new boolean fields (Recidiviz/recidiviz-data#31865)
## Description of the change

This fixes this error in today's alpha deploy:
```
>  >  ERROR [root] Migrations failed to run: (psycopg2.errors.NotNullViolation) column "has_dismissed_data_unavailable_note" of relation "user_metadata" contains null values
```

by setting a server default for the new columns. Since the migration
step failed outright, it seemed like deleting and recreating the
migration file was the way to go. I also ran the admin panel container
locally and the migration did correctly add the new columns and set
their data type to booleans with a false default.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e41e507bef5d7ab24692b614965648d05b9fd089