[Testing] IngestViewEmulatorQueryTestCase state_code and ingest_view_name properties (Recidiviz/recidiviz-data#31867)
I will be encapsulating and using some raw data fixture logic of the
`IngestViewEmulatorQueryTestCase` for some integration tests and needed
to use `StateCode`. This PR updates all ingest view tests to use the
state code enum instead of the region code string. I also addressed the
TODO to have the parent class get the ingest view, rather than the test
setUp grab it.

I added a `from_state_code` method to
`DirectIngestViewQueryBuilderCollector` to allow us to instantiate it
from a `StateCode` as well, which will also be helpful in future
integration tests.

GitOrigin-RevId: 9024663f53e2ffeaec9237b2e928dd2724b9fa1b