Adds new view for live workflows by state (Recidiviz/recidiviz-data#31866)
## Description of the change

Adds a new view that uses tool usage information to identify the set of
task types that are live in each state. This view also joins to a
materialized version of the `workflows_launch_metadata` table (see
Recidiviz/recidiviz-data#31841), which contains all fully launched opportunities, so that we can
pull in a flag indicating whether a full launch has occurred for that
completion event type in a particular state.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Advances Recidiviz/recidiviz-data#31645

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 56c047cd6cd931f9bc94636fc15837b5f7deb7c2