Rename supervision legal authority sessions (Recidiviz/recidiviz-data#31800)
## Description of the change

This PR is stacked on Recidiviz/recidiviz-data#31390. `supervision_legal_authority_session`s was
created (and named such) with the idea that it would eventually be
sourced based on the `legal_authority field`. However, the legal
authority concept is no longer going to be introduced, and as such we
have decided to instead name this view
`prioritized_supervision_sessions`

Downstream references were tested with the following load views to
sandbox command:

```
python -m recidiviz.tools.load_views_to_sandbox \
        --sandbox_dataset_prefix agaidus2 --prompt manual \
        --view_ids_to_load sessions.compartment_sub_sessions_preprocessed,sessions.compartment_sub_sessions,sessions.prioritized_supervision_sessions,analyst_data.fines_fees_sessions,analyst_data.us_tn_fines_fees_sessions_preprocessed,sessions.supervision_officer_transitional_caseload_type_sessions \
        --update_ancestors False \
        --update_descendants False
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9bcf8ee6672a3e87c737773da23f5030f632e4de