Implement PSI Case Insights recidivism calculation rollup (Recidiviz/recidiviz-data#31655)
## Description of the change

This PR implements rollup for the calculation of Case Insights
recidivism rates.

The rollup levels, described in Recidiviz/recidiviz-data#29745, are:
1. Specific charge + LSI-R score bucket (24% of all historical offenses)
2. Specific charge (50%)
3. NCIC category + LSI-R score bucket (5%)
4. NCIC category (7%)
5. Combined offense category (12%) (see below)
6. Violent vs non-violent (0.8%)
7. All offenses (0%)

The rollup procedure is as follows:
- The recidivism rates for each sentence disposition (PROBATION, RIDER,
TERM) are calculated for each level. For each disposition, the rates are
calculated for a series of months up to 36 months, and confidence
intervals (CIs) are calculated at each point on the curve.
- A level is considered qualified if the CI size of the final month is
less than or equal to `ROLLUP_CI_THRESHOLD` (set to 0.2) for all three
dispositions. Note that if there is no data for a given disposition,
then it isn't considered to violate the threshold because a recidivism
curve for that disposition won't be plotted at all.
- The lowest (finest-grained) level that meets this qualification is
chosen as the rollup level.
- The corresponding recidivism rates are recorded in
`recidivism_probation_series`, `recidivism_rider_series`,
`recidivism_term_series`.
- The sum of the cohort sizes for the three dispositions in the chosen
cohort level is recorded in `recidivism_num_records`.
- The attributes of the chosen rollup level (e.g. {"state_code",
"most_severe_description"}) are recorded in the `recidivism_rollup`.

Note that the disposition fields of the Case Insights table (the
percentage of people sentenced to each disposition) are still calculated
at the most granular level and aren't subject to this rollup logic.

This PR only includes rollup levels 1-4 and 7, and levels 5 and 6 will
be added in a follow-up PR.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Advances Recidiviz/recidiviz-data#30755

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 479e7108710f1daa911078b4c2a3c630bdb2d276