Change CI calculation method for Case Insights to 'beta' (Recidiviz/recidiviz-data#31771)
## Description of the change

We previously used the standard method (approximation using a normal
distribution) for confidence interval (CI) calculation, which is
inexact, and in fact is wildly incorrect with a small number of data
points.

In particular, when there is only one data point in the sample, the
normal-approximated CI size is 0. (That is, using our recidivism
scenario: either the one person in the sample recidivated, in which case
the recidivism rate is 100% and the approximated CI is [100%, 100%]; or
they didn't recidivate, in which case the rate is 0% and the
approximated interval is [0%, 0%].) This is completely incorrect (the
actual exact confidence size with one data point is ~95%), and is
especially problematic for our setting [once we implement
roll-up](https://app.zenhub.com/workspaces/sentencing-6542cc4d7b52e505eb78b52b/issues/gh/recidiviz/recidiviz-data/30755),
which relies on the CI size -- this would mean we would display
recidivism curves for populations with one example, when instead we
should be rolling up to cohorts with larger population sizes.

This PR switches to using the Clopper-Pearson approximation (using
[statsmodels.stats.proportion.proportion_confint](https://www.statsmodels.org/dev/generated/statsmodels.stats.proportion.proportion_confint.html)
with `method='beta'`), which is a much better approximation especially
with a small number of samples (e.g. with one sample the CI size is
97.5%). We could use the 'binom_test' method of `proportion_confint`
which calculates the exactly correct CI, but it's computationally
expensive for cases with a large number of data points (the whole
calculation job runs in ~6 min compared to ~15 seconds) and the
differences between it and Clopper-Pearson are minimal.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#31287

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8827b32db43953ebe22badbd970bb7a851423f5d