Add roll-up levels based on offense categories for Case Insights recidivism calculations (Recidiviz/recidiviz-data#31869)
## Description of the change

As a follow-up to Recidiviz/recidiviz-data#31655, this PR adds roll-up levels 5 and 6:
5. Combined offense category
6. Violent vs non-violent

These are based on the `any_is_violent_uniform`, `any_is_drug_uniform`,
and `any_is_sex_offense` columns of the cohort DataFrame, which are
derived from `sessions.sentence_imposed_group_summary_materialized`.

The `combined_offense_category` is a string that represents whether the
offense is a violent offense, drug offense, and/or sex offense. It can
be more than one, in which case it's a comma-separated string (e.g.
"Violent offense, Drug offense"), or just one (e.g. "Drug offense"), or
none (in which case it's "Non-drug, Non-violent, Non-sex offense").

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#30755

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 95d3d80c5b881fb709b24d38506359d272098078