[raw data import dag] integration tests for resource locking & step 2 (Recidiviz/recidiviz-data#31767)
## Description of the change

adds a few integration tests for resource locking behavior for as well
as an initial test for step 2 (is cool to watch everything run)!

the test for step 2 sets a mock input for what we might see in the
bucket and validates that the xcom output of the
`split_by_pre_import_normalization_type` task matches what we expect it
to.

since we can't pickle modules, we cant set the `region_module_override`
before `.execute` or `.function` call time, so I had to do a little
dance to get `region_module_override` properly set

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#31755

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0ab72a2f39185613d76a3ca6711972fa736833b8