[Justice Counts] Automate user permissions check (Recidiviz/recidiviz-data#31852)
## Description of the change

Automating the user permissions check so that we don't have to check the
spreadsheet! Instead of writing to the spreadsheet, this job will post
to #jc-tech-chat so CSG can handle the results as they see fit.

It will flag:
* CSG users with non-read-only access on prod
* Users with access to superagency but not all of its children
* Users in > 10 non-child agencies
* Users in 2+ agencies but no superagency
* Users in agencies in multiple states

However, compared to our old check, it no longer flags the following:
* CSG users missing access on staging or prod
* My reasoning is: if a CSG user notices they are missing access to an
agency they need, they can just use the Admin Panel to grant themselves
access
* Non-Rec/CSG users with access to staging
* My reasoning is: there were enough exceptions to this one that it felt
hard to figure out how to include in a sustainable way. Also, if this
happens, it's not the end of the world, since there isn't real data on
staging.

How I tested: A bajillion messages to jc-eng-only. Sorry, guys!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f95871b7548a2d5cf090421bbf943c20b5d6cbff